requirements.txt
certifi
charset-normalizer
click
colorama
Flask
idna
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
python-dateutil
pytz
requests
six
urllib3
Werkzeug
flask-mysqldb
mysqlclient
python-dotenv
pymysql
mysql-connector-python
